requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiofiles==23.2.1
aiohttp==3.9.5